requirements.txt
torch==2.1
transformers==4.37.2
ctransformers
gradio
fastapi
python-dotenv
langchain
chromadb
pypdf
sentence-transformers
tiktoken
tqdm
together
langchain-together
langchain-community
langchain-google-genai
logging
faiss-cpu
streamlit
utils
pinecone-client
langchain-pinecone
bs4
fastapi
uvicorn
beautifulsoup4
requests